+++ navigation/src/main/kotlin/com/imashnake/animite/navigation/NavigationBar.kt
@@ -1,6 +1,5 @@
 package com.imashnake.animite.navigation
 
-import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.WindowInsets
@@ -9,7 +8,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.selection.selectableGroup
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBarDefaults
 import androidx.compose.material3.NavigationBarItem
@@ -21,17 +19,10 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
-import androidx.navigation.NavDestination.Companion.hasRoute
-import androidx.navigation.NavDestination.Companion.hierarchy
-import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.compose.currentBackStackEntryAsState
 
 @Composable
@@ -74,65 +65,4 @@ fun NavigationBar(
             }
         }
     }
-}
-
-enum class NavigationBarPaths(
-    val navigateTo: (NavController) -> Unit,
-    val matchesDestination: (NavBackStackEntry) -> Boolean,
-    val icon: @Composable () -> Unit,
-    @StringRes val labelRes: Int
-) {
-    Social(
-        navigateTo = {
-            it.navigate(SocialRoute) {
-                popUpTo(id = it.graph.findStartDestination().id) {
-                    saveState = true
-                }
-                launchSingleTop = true
-            }
-        },
-        matchesDestination = {
-            it.destination.hierarchy.any { it.hasRoute(SocialRoute::class) }
-        },
-        icon = {
-            Icon(ImageVector.vectorResource(R.drawable.social), contentDescription = stringResource(R.string.social))
-        },
-        labelRes = R.string.social
-    ),
-    Home(
-        navigateTo = {
-            it.navigate(HomeRoute) {
-                popUpTo(id = it.graph.findStartDestination().id) {
-                    saveState = true
-                    inclusive = true
-                }
-                launchSingleTop = true
-            }
-        },
-        matchesDestination = {
-            it.destination.hierarchy.any { it.hasRoute(HomeRoute::class) }
-        },
-        icon = {
-            Icon(ImageVector.vectorResource(R.drawable.home), contentDescription = stringResource(R.string.home))
-        },
-        labelRes = R.string.home
-    ),
-
-    Profile(
-        navigateTo = {
-            it.navigate(ProfileRoute()) {
-                popUpTo(id = it.graph.findStartDestination().id) {
-                    saveState = true
-                }
-                launchSingleTop = true
-            }
-        },
-        matchesDestination = {
-            it.destination.hierarchy.any { it.hasRoute(ProfileRoute::class) }
-        },
-        icon = {
-            Icon(ImageVector.vectorResource(R.drawable.profile), contentDescription = stringResource(R.string.profile))
-        },
-        labelRes = R.string.profile
-    ),
 }
